Renamed some packages to fallow naming conventions. Removed unused imports. Update suite with correct package names
