+++ src/main/java/com/revature/cmsforce/CMSforceApplication.java
@@ -25,8 +25,11 @@ public class CMSforceApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(CMSforceApplication.class, args);
 		
-	
 		
 	}
 
+}
+
+interface Test{
+	
 }
