+++ src/test/java/com/revature/entities/LinkTest.java
@@ -1,4 +1,4 @@
-package com.revature.entitiestests;
+package com.revature.entities;
 
 import static org.testng.Assert.assertTrue;
