+++ src/test/java/com/revature/entities/ReqLinkTest.java
@@ -1,4 +1,4 @@
-package com.revature.entitiestests;
+package com.revature.entities;
 
 import static org.testng.Assert.assertTrue;
 import org.testng.annotations.Test;
