+++ src/test/java/com/revature/entities/RequestTest.java
@@ -1,4 +1,4 @@
-package com.revature.entitiestests;
+package com.revature.entities;
 
 import org.testng.annotations.Test;
 import org.testng.annotations.BeforeTest;
