+++ src/test/java/com/revature/suite/testNGsuite.xml
@@ -11,17 +11,17 @@
 			<class name = "com.revature.controllertests.MetricsControllerTest"/>
 			<class name = "com.revature.controllertests.LinkControllerTest"/>
 			<!-- These are the entities tests -->
-			<class name="com.revature.entitiestests.ContentTest" />
-			<class name="com.revature.entitiestests.LinkTest" />
-			<class name="com.revature.entitiestests.ModuleTest" />
-			<class name="com.revature.entitiestests.ReqLinkTest" />
-			<class name="com.revature.entitiestests.RequestTest" />
+			<class name="com.revature.entities.ContentTest" />
+			<class name="com.revature.entities.LinkTest" />
+			<class name="com.revature.entities.ModuleTest" />
+			<class name="com.revature.entities.ReqLinkTest" />
+			<class name="com.revature.entities.RequestTest" />
 			<!-- These are the JSONEntities tests -->
-			<class name="com.revature.JSONEntitiestests.JSONContentTest" />
-			<class name="com.revature.JSONEntitiestests.JSONModuleTest" />
-			<class name="com.revature.JSONEntitiestests.JSONRequestTest" />
+			<class name="com.revature.JSONEntities.JSONContentTest" />
+			<class name="com.revature.JSONEntities.JSONModuleTest" />
+			<class name="com.revature.JSONEntities.JSONRequestTest" />
 			<!-- These are the cmsforce tests -->
-			<class name="com.revature.cmsforcetests.CMSforceApplicationTest" />
+			<class name="com.revature.cmsforce.CMSforceApplicationTest" />
 			<!-- These are the Exception tests -->
 			<class name="com.revature.exceptions.InvalidContentExceptionTest" />
 			<class name="com.revature.exceptions.InvalidContentIdTest" />
@@ -35,12 +35,12 @@
 			<class name="com.revature.servicestests.ModuleServiceTest" />
 			<class name="com.revature.servicestests.SearchServiceTest" />
 			<!-- These are the Util tests -->
-			<class name = "com.revature.utiltests.MetricsDataTest"/>
-			<class name = "com.revature.utiltests.TimeGraphTest"/>
-			<class name = "com.revature.utiltests.ValidationUtilTest"/>
-			<class name = "com.revature.utiltests.ControllerLogAspectTest"/>
-			<class name = "com.revature.utiltests.SeviceLogAspectTest"/>
-			<class name = "com.revature.utiltests.ExceptionAspectTest"/>	
+			<class name = "com.revature.util.MetricsDataTest"/>
+			<class name = "com.revature.util.TimeGraphTest"/>
+			<class name = "com.revature.util.ValidationUtilTest"/>
+			<class name = "com.revature.util.ControllerLogAspectTest"/>
+			<class name = "com.revature.util.SeviceLogAspectTest"/>
+			<class name = "com.revature.util.ExceptionAspectTest"/>	
 		</classes>
 	</test>
 </suite>
